rock-classifier-app/backend/requirements.txt
# Web Framework & Server
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6

# Deep Learning & Image Processing
torch==2.2.0
torchvision==0.17.0
Pillow==10.1.0

# Database
SQLAlchemy==2.0.23

# Utilities
python-dotenv==1.0.0
aiofiles==23.2.1
typing-extensions==4.8.0

# Development & Testing
pytest==7.4.3
pytest-asyncio==0.21.1
